+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <a name="1.91"></a>
 
+- Include HOPR Staking Season 5 smart contracts ([#4221](https://github.com/hoprnet/hoprnet/pull/4221))
+
 ---
 
 <a name="1.90"></a>